main.swift -> main2.swift & flowchart_dongdong.jpg -> Step2_flowchart_dongdong.jpg
